+++ src/compat/db_query.c
@@ -266,19 +266,19 @@ or_pack_query_format (char *buf, const DB_QUERY_TYPE * columns, const int count)
       /* column type */
       ptr = or_pack_int (ptr, (int) column->col_type);
       /* column name */
-      len = (column->name == NULL) ? 0 : strlen (column->name);
+      len = (column->name == NULL) ? 0 : (int) strlen (column->name);
       ptr = or_pack_string_with_length (ptr, column->name, len);
 
       /* attribute name */
-      len = (column->attr_name == NULL) ? 0 : strlen (column->attr_name);
+      len = (column->attr_name == NULL) ? 0 : (int) strlen (column->attr_name);
       ptr = or_pack_string_with_length (ptr, column->attr_name, len);
 
       /* spec name */
-      len = (column->spec_name == NULL) ? 0 : strlen (column->spec_name);
+      len = (column->spec_name == NULL) ? 0 : (int) strlen (column->spec_name);
       ptr = or_pack_string_with_length (ptr, column->spec_name, len);
 
       /* user specified column name */
-      len = (column->original_name == NULL) ? 0 : strlen (column->original_name);
+      len = (column->original_name == NULL) ? 0 : (int) strlen (column->original_name);
       ptr = or_pack_string_with_length (ptr, column->original_name, len);
       /* column data type */
       ptr = or_pack_int (ptr, column->db_type);
